refactor: move resources data into the resources folder
