+++ carrerait-roadmap/src/data/resources/resources.ts
